+++ runtimes/runtimes/lsp/router/lspRouter.ts
@@ -22,17 +22,26 @@ import {
     UpdateConfigurationParams,
     updateConfigurationRequestType,
     HandlerResult,
+    WorkspaceFolder,
+    DidChangeWorkspaceFoldersParams,
+    WorkspaceFoldersChangeEvent,
+    CreateFilesParams,
+    DeleteFilesParams,
+    RenameFilesParams,
+    DidSaveTextDocumentParams,
 } from '../../../protocol'
 import { Connection } from 'vscode-languageserver/node'
 import { LspServer } from './lspServer'
 import { findDuplicates, mergeObjects } from './util'
 import { CredentialsType, PartialInitializeResult } from '../../../server-interface'
 import { SERVER_CAPABILITES_CONFIGURATION_SECTION } from './constants'
+import { getWorkspaceFoldersFromInit } from './initializeUtils'
 
 export class LspRouter {
     private initializeResult?: InitializeResult
     public clientInitializeParams?: InitializeParams
     public servers: LspServer[] = []
+    private workspaceFolders: WorkspaceFolder[] = []
 
     constructor(
         private lspConnection: Connection,
@@ -46,6 +55,10 @@ export class LspRouter {
         lspConnection.onRequest(getConfigurationFromServerRequestType, this.getConfigurationFromServer)
         lspConnection.onNotification(notificationFollowupRequestType, this.onNotificationFollowup)
         lspConnection.onRequest(updateConfigurationRequestType, this.updateConfiguration)
+        lspConnection.workspace.onDidCreateFiles(this.didCreateFiles)
+        lspConnection.workspace.onDidDeleteFiles(this.didDeleteFiles)
+        lspConnection.workspace.onDidRenameFiles(this.didRenameFiles)
+        lspConnection.onDidSaveTextDocument(this.didSaveTextDocument)
     }
 
     initialize = async (
@@ -71,6 +84,12 @@ export class LspRouter {
             this.lspConnection.console.log(
                 `Unknown initialization error\nwith initialization options: ${JSON.stringify(params.initializationOptions)}`
             )
+        }
+
+        this.workspaceFolders = getWorkspaceFoldersFromInit(this.lspConnection.console, params)
+
+        if (this.workspaceFolders.length === 0) {
+            this.lspConnection.console.info('No workspace folders found in initialization parameters')
         }
 
         let responsesList = await Promise.all(this.servers.map(s => s.initialize(params, token)))
@@ -114,6 +133,33 @@ ${JSON.stringify({ ...result.capabilities, ...result.awsServerCapabilities })}`
         )
 
         return result
+    }
+
+    didChangeWorkspaceFolders = (event: WorkspaceFoldersChangeEvent): void => {
+        const supportsWorkspaceFolders = this.clientInitializeParams?.capabilities.workspace?.workspaceFolders === true
+
+        if (!supportsWorkspaceFolders) {
+            this.lspConnection.console.warn(
+                "Client doesn't support sending workspace folder change events. Ignoring workspace folder changes."
+            )
+            return
+        }
+
+        this.workspaceFolders = this.workspaceFolders.filter(
+            folder => !event.removed.some(removed => removed.uri === folder.uri)
+        )
+        this.workspaceFolders.push(...event.added)
+        const params: DidChangeWorkspaceFoldersParams = { event }
+
+        this.routeNotificationToAllServers((server, params) => {
+            if (server.sendDidChangeWorkspaceFoldersNotification) {
+                server.sendDidChangeWorkspaceFoldersNotification(params)
+            }
+        }, params)
+    }
+
+    getAllWorkspaceFolders(): WorkspaceFolder[] {
+        return this.workspaceFolders
     }
 
     executeCommand = async (
@@ -179,11 +225,35 @@ ${JSON.stringify({ ...result.capabilities, ...result.awsServerCapabilities })}`
             this.lspConnection.client.register(DidChangeConfigurationNotification.type, undefined)
         }
 
+        // Register for workspace folder changes only if supported
+        if (workspaceCapabilities?.workspaceFolders === true) {
+            this.lspConnection.workspace.onDidChangeWorkspaceFolders(this.didChangeWorkspaceFolders)
+        }
+
         this.routeNotificationToAllServers((server, params) => server.sendInitializedNotification(params), params)
     }
 
     onNotificationFollowup = (params: NotificationFollowupParams): void => {
         this.routeNotificationToAllServers((server, params) => server.sendNotificationFollowup(params), params)
+    }
+
+    didCreateFiles = (params: CreateFilesParams): void => {
+        this.routeNotificationToAllServers((server, params) => server.sendDidCreateFilesNotification(params), params)
+    }
+
+    didDeleteFiles = (params: DeleteFilesParams): void => {
+        this.routeNotificationToAllServers((server, params) => server.sendDidDeleteFilesNotification(params), params)
+    }
+
+    didRenameFiles = (params: RenameFilesParams): void => {
+        this.routeNotificationToAllServers((server, params) => server.sendDidRenameFilesNotification(params), params)
+    }
+
+    didSaveTextDocument = (params: DidSaveTextDocumentParams): void => {
+        this.routeNotificationToAllServers(
+            (server, params) => server.sendDidSaveTextDocumentNotification(params),
+            params
+        )
     }
 
     private async routeRequestToFirstCapableServer<P, R>(